+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.6] Q2 2023
+ - factory-reset: fixed reporting "in_progress"
+ - network-status: fixed updating status after network adapters changed
+
 ## [0.11.5] Q2 2023
  - systemd: removed enforcing reboot on errors